[web] Add device-utils
- Provides DeviceSelector and DeviceList.
- Add isMultiple to DeviceSelector.
