+++ web/src/components/storage/index.js
@@ -30,5 +30,5 @@ export { default as DASDFormatProgress } from "./DASDFormatProgress";
 export { default as ZFCPPage } from "./ZFCPPage";
 export { default as ZFCPDiskForm } from "./ZFCPDiskForm";
 export { default as ISCSIPage } from "./ISCSIPage";
-export { default as DeviceSelector } from "./DeviceSelector";
+export { DeviceList, DeviceSelector } from "./device-utils";
 export { default as VolumeForm } from "./VolumeForm";
